Move a bunch of methods
