+++ src/main/kotlin/propCheck/arbitrary/HelperClasses.kt
