+++ src/test/kotlin/propCheck/PropertySpec.kt
@@ -3,6 +3,7 @@ package propCheck
 import io.kotlintest.AbstractSpec
 import io.kotlintest.TestType
 import io.kotlintest.specs.IntelliMarker
+import propCheck.property.Property
 
 abstract class AbstractPropertySpec(f: AbstractPropertySpec.() -> Unit = {}) : AbstractSpec() {
     init { f() }
